Split editor feature-wall animation
